refactor all pages removing Banner and putting Banner in SocialTiger
