+++ api/database.json
@@ -23,9 +23,9 @@
       "userId": 2
     },
     {
+      "id": 4,
       "folderName": "Mommy",
-      "userId": 3,
-      "id": 4
+      "userId": 3
     },
     {
       "folderName": "Liam",
@@ -54,13 +54,13 @@
     },
     {
       "id": 13,
-      "folderName": "Everyday BGs!!!",
+      "folderName": "Everyday BGs!",
       "userId": 1
     },
     {
-      "folderName": "Holiday Fun",
-      "userId": 1,
-      "id": 14
+      "id": 14,
+      "folderName": "Holiday",
+      "userId": 1
     },
     {
       "folderName": "Social Tiger",
@@ -73,11 +73,6 @@
       "userId": 1
     },
     {
-      "folderName": "folderList test",
-      "userId": 1,
-      "id": 20
-    },
-    {
       "id": 21,
       "folderName": "bob456",
       "userId": 2
@@ -86,11 +81,6 @@
       "id": 22,
       "folderName": "65432",
       "userId": 2
-    },
-    {
-      "folderName": "folder6",
-      "userId": 1,
-      "id": 23
     },
     {
       "folderName": "Flowers",
@@ -156,11 +146,11 @@
       "id": 32
     },
     {
-      "imageName": "Spiderman 2",
-      "imageDescription": "Spiderman 2",
+      "id": 33,
+      "imageName": "Spiderman Skyline",
+      "imageDescription": "Spiderman skyline image ",
       "url": "https://res.cloudinary.com/lordargyle22/image/upload/v1596813187/socialTiger/spiderman_logo_vxhlxb.png",
-      "folderId": 7,
-      "id": 33
+      "folderId": "7"
     },
     {
       "imageName": "Mother's Day",
@@ -420,6 +410,13 @@
       "url": "https://res.cloudinary.com/lordargyle22/image/upload/v1597285748/socialTiger/WildflowerField_a3n149.jpg",
       "folderId": 24,
       "id": 73
+    },
+    {
+      "imageName": "Planet Hulk",
+      "imageDescription": "Hulk from Thor: Ragnarok",
+      "url": "https://res.cloudinary.com/lordargyle22/image/upload/v1597287370/socialTiger/PlanetHulk_yo2qxv.jpg",
+      "folderId": 6,
+      "id": 74
     }
   ]
 }
