+++ src/components/Banner.js
@@ -1,13 +1,13 @@
 import React from "react";
-import BannerForLogin from "../components/BannerForLogin";
+// import BannerForLogin from "../components/BannerForLogin";
 import BannerSignedIn from "../components/BannerSignedIn";
 import "./Banner.css"
 
-const isAuthenticated = () => sessionStorage.getItem("credentials") !== null;
+// const isAuthenticated = () => sessionStorage.getItem("credentials") !== null;
 
 const Banner = (props) => {
 
-    if (isAuthenticated()) {
+    // if (isAuthenticated()) {
 
         return( 
             <>
@@ -16,15 +16,15 @@ const Banner = (props) => {
                 </div>
             </>
         )
-    } else {
-        return (
-            <>
-                <div className="bannerForLogin">
-                    <BannerForLogin />
-                </div>
-            </>
-        )
-    }
+    // } else {
+    //     return (
+    //         <>
+    //             <div className="bannerForLogin">
+    //                 <BannerForLogin />
+    //             </div>
+    //         </>
+    //     )
+    // }
 };
 
 export default Banner;
