+++ src/components/BannerForLogin.js
@@ -1,22 +1,22 @@
-import React from "react";
+// import React from "react";
 
 
-const BannerForLogin = (props) => {
+// const BannerForLogin = (props) => {
 
-    return (
-        <>
-                <div className="h1-container">
-                    <h1>SOCIAL TIGER</h1>
-                </div>
-                <div>
+//     return (
+//         <>
+//                 <div className="h1-container">
+//                     <h1>SOCIAL TIGER</h1>
+//                 </div>
+//                 <div>
                     
-                    <img 
-                        className="logo" 
-                        src="https://res.cloudinary.com/lordargyle22/image/upload/v1596292629/socialTiger/SocialTiger_bwtvfg.png" />
-                </div> 
+//                     <img 
+//                         className="logo" 
+//                         src="https://res.cloudinary.com/lordargyle22/image/upload/v1596292629/socialTiger/SocialTiger_bwtvfg.png" />
+//                 </div> 
             
-        </>
-    );
-};
+//         </>
+//     );
+// };
 
-export default BannerForLogin;
+// export default BannerForLogin;
