+++ src/components/nav/NavBar.css
@@ -1,4 +1,12 @@
 
+  button.logOutButton {
+    position: absolute;
+    top: 15px;
+    left: 15px;
+    height: auto;
+    font-size: 18px;
+}
+  
   nav .container {
     display: flex;
     justify-content: center;
@@ -6,6 +14,7 @@
     padding: 0;
     margin-bottom: 0;
   }
+
   
   nav .container li {
     flex: 1;
