+++ src/pages/CreateFolderPage.js
@@ -1,14 +1,14 @@
 import React from "react";
 import NavBar from "../components/nav/NavBar"
 import Authentication from "../components/auth/Authentication";
-import Banner from "../components/Banner"
+// import Banner from "../components/Banner"
 import FolderForm from "../components/folder/FolderForm"
 
 const CreateFolderPage = (props) => {
 
     return (
         <>
-            <Banner/>
+            {/* <Banner/> */}
             <NavBar/>
             <div>
                 <h3>Create Folder</h3>
