+++ src/pages/CreateImagePage.css
@@ -9,7 +9,3 @@ div.underNav {
 h3 {
     color: yellowgreen
 }
-
-{
-
-}
