+++ src/pages/CreateImagePage.js
@@ -1,7 +1,7 @@
 import React from "react";
 import NavBar from "../components/nav/NavBar"
 import Authentication from "../components/auth/Authentication";
-import Banner from "../components/Banner"
+// import Banner from "../components/Banner"
 import ImageForm from "../components/image/ImageForm"
 import "./CreateImagePage.css"
 
@@ -10,7 +10,7 @@ const CreateImagePage = (props) => {
 
     return (
         <>
-            <Banner/>
+            {/* <Banner/> */}
             <NavBar/>
             <div className="underNav">
                 {/* <h3>Create Image</h3> */}
