+++ src/pages/EditFolderPage.js
@@ -1,13 +1,13 @@
 import React from "react";
 import NavBar from "../components/nav/NavBar";
 import Authentication from "../components/auth/Authentication";
-import Banner from "../components/Banner";
+// import Banner from "../components/Banner";
 import FolderEditForm from "../components/folder/FolderEditForm";
 
 const EditFolderPage = (props) => {
     return (
         <>
-            <Banner />
+            {/* <Banner /> */}
             <NavBar />
             <div>
                 <h3>Edit Folder</h3>
