+++ src/pages/EditImagePage.js
@@ -1,13 +1,13 @@
 import React from "react";
 import NavBar from "../components/nav/NavBar";
 import Authentication from "../components/auth/Authentication";
-import Banner from "../components/Banner";
+// import Banner from "../components/Banner";
 import ImageEditForm from "../components/image/ImageEditForm";
 
 const EditImagePage = (props) => {
     return (
         <>
-            <Banner />
+            {/* <Banner /> */}
             <NavBar />
             <div>
                 <h3>Edit Image</h3>
