+++ src/pages/FolderListPage.js
@@ -1,14 +1,14 @@
 import React from "react";
 import NavBar from "../components/nav/NavBar";
 import Authentication from "../components/auth/Authentication";
-import Banner from "../components/Banner";
+// import Banner from "../components/Banner";
 import FolderList from "../components/folder/FolderList"
 import "../components/folder/FolderList.css"
 
 const FolderListPage = (props) => {
     return (
         <>
-            <Banner />
+            {/* <Banner /> */}
             <NavBar/>
             <div className="folderListBody">
                 <h3>My Folders</h3>
