+++ src/pages/HomePage.js
@@ -1,7 +1,7 @@
 import React from "react";
-import NavBar from "../components/nav/NavBar"
 import Authentication from "../components/auth/Authentication";
-import Banner from "../components/Banner"
+import NavBar from "../components/nav/NavBar"
+// import Banner from "../components/Banner"
 import ImageRandomizer from "../components/image/ImageRandomizer"
 import Welcome from "../components/Welcome"
 
@@ -9,7 +9,7 @@ const HomePage = (props) => {
 
   return (
       <>
-        <Banner/>
+        {/* <Banner/> */}
         <NavBar/>
         <div className="homePageBody">
           <Welcome
