+++ src/pages/ImageListPage.js
@@ -1,13 +1,13 @@
 import React from "react";
 import NavBar from "../components/nav/NavBar";
 import Authentication from "../components/auth/Authentication";
-import Banner from "../components/Banner";
+// import Banner from "../components/Banner";
 import ImageList from "../components/image/ImageList"
 
 const ImageListPage = (props) => {
     return (
         <>
-            <Banner />
+            {/* <Banner /> */}
             <NavBar/>
             <div className="imageListBody">
                 <h3>My Images</h3>
