+++ src/pages/ImageViewPage.js
@@ -1,13 +1,13 @@
 import React from "react";
 import NavBar from "../components/nav/NavBar";
 import Authentication from "../components/auth/Authentication";
-import Banner from "../components/Banner";
+// import Banner from "../components/Banner";
 import ImageView from "../components/image/ImageView";
 
 const ImageViewPage = (props) => {
     return (
         <>
-            <Banner />
+            {/* <Banner /> */}
             <NavBar/>
             <div className="imageViewBody">
                 <ImageView imageId={parseInt(props.match.params.imageId)} {...props} />
