+++ src/pages/LoginPage.js
@@ -1,13 +1,13 @@
 import React from "react";
 import Login from "../components/auth/Login";
 import {withRouter} from "react-router-dom";
-import Banner from "../components/Banner";
+// import Banner from "../components/Banner";
 
 const LoginPage = (props) => {
 
     return (
         <>
-            <Banner/>
+            {/* <Banner />  */}
             <Login { ...props}/>
         </>
     )
