+++ src/pages/RegistrationPage.js
@@ -1,13 +1,13 @@
 import React from "react";
 import Registration from "../components/auth/RegisterNewUser";
 import {withRouter} from "react-router-dom";
-import Banner from "../components/Banner";
+// import Banner from "../components/Banner";
 
 const RegistrationPage = (props) => {
 
     return (
         <>
-            <Banner/>
+            {/* <Banner/> */}
             <Registration { ...props}/>
         </>
     )
